Move `OrderEntity` to correct package `org.wordpress.android.fluxc.persistence.entity`
