+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/wear/extensions/OrderEntityExt.kt
@@ -3,7 +3,7 @@ package com.woocommerce.android.wear.extensions
 import com.woocommerce.commons.WearOrder
 import com.woocommerce.commons.WearOrderAddress
 import org.wordpress.android.fluxc.model.LocalOrRemoteId
-import org.wordpress.android.fluxc.model.OrderEntity
+import org.wordpress.android.fluxc.persistence.entity.OrderEntity
 import org.wordpress.android.fluxc.model.order.OrderAddress
 
 fun OrderEntity.toWearOrder() = WearOrder(
