+++ WooCommerce-Wear/src/test/java/com/woocommerce/android/wear/ui/orders/list/FetchOrdersTest.kt
@@ -20,7 +20,7 @@ import org.junit.Test
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.whenever
 import org.wordpress.android.fluxc.model.LocalOrRemoteId
-import org.wordpress.android.fluxc.model.OrderEntity
+import org.wordpress.android.fluxc.persistence.entity.OrderEntity
 import org.wordpress.android.fluxc.model.SiteModel
 import org.wordpress.android.fluxc.store.WCWearableStore.OrdersForWearablesResult.Success
