+++ libs/fluxc-plugin/src/main/kotlin/org/wordpress/android/fluxc/persistence/entity/OrderEntity.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.fluxc.model
+package org.wordpress.android.fluxc.persistence.entity
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
