+++ libs/fluxc-plugin/src/main/kotlin/org/wordpress/android/fluxc/store/WCWearableStore.kt
@@ -1,6 +1,6 @@
 package org.wordpress.android.fluxc.store
 
-import org.wordpress.android.fluxc.model.OrderEntity
+import org.wordpress.android.fluxc.persistence.entity.OrderEntity
 import org.wordpress.android.fluxc.model.SiteModel
 import org.wordpress.android.fluxc.network.rest.wpcom.auth.AccessToken
 import org.wordpress.android.fluxc.network.rest.wpcom.wc.order.OrderRestClient
